refactor(cli): split plan-mode-toggle.ts into 7 focused modules
