+++ packages/cli/src/extensions/plan-mode/todo-items.ts
@@ -1,0 +1,73 @@
+import type { AgentMessage } from "@mariozechner/pi-agent-core";
+import type { AssistantMessage, TextContent } from "@mariozechner/pi-ai";
+
+// ── Todo item types ──────────────────────────────────────────────────────────
+
+export interface PlanTodoItem {
+    step: number;
+    text: string;
+    completed: boolean;
+}
+
+export function cleanStepText(text: string): string {
+    let cleaned = text
+        .replace(/\*{1,2}([^*]+)\*{1,2}/g, "$1")
+        .replace(/`([^`]+)`/g, "$1")
+        .replace(/^(Use|Run|Execute|Create|Write|Read|Check|Verify|Update|Modify|Add|Remove|Delete|Install)\s+(the\s+)?/i, "")
+        .replace(/\s+/g, " ")
+        .trim();
+    if (cleaned.length > 0) cleaned = cleaned.charAt(0).toUpperCase() + cleaned.slice(1);
+    if (cleaned.length > 60) cleaned = `${cleaned.slice(0, 57)}...`;
+    return cleaned;
+}
+
+export function extractTodoItems(message: string): PlanTodoItem[] {
+    const items: PlanTodoItem[] = [];
+    const headerMatch = message.match(/\*{0,2}Plan:\*{0,2}\s*\n/i);
+    if (!headerMatch) return items;
+
+    const planSection = message.slice(message.indexOf(headerMatch[0]) + headerMatch[0].length);
+    const numberedPattern = /^\s*(\d+)[.)]\s+\*{0,2}([^*\n]+)/gm;
+
+    for (const match of planSection.matchAll(numberedPattern)) {
+        const text = match[2].trim().replace(/\*{1,2}$/, "").trim();
+        if (text.length > 5 && !text.startsWith("`") && !text.startsWith("/") && !text.startsWith("-")) {
+            const cleaned = cleanStepText(text);
+            if (cleaned.length > 3) {
+                items.push({ step: items.length + 1, text: cleaned, completed: false });
+            }
+        }
+    }
+    return items;
+}
+
+export function extractDoneSteps(message: string): number[] {
+    const steps: number[] = [];
+    for (const match of message.matchAll(/\[DONE:(\d+)\]/gi)) {
+        const step = Number(match[1]);
+        if (Number.isFinite(step)) steps.push(step);
+    }
+    return steps;
+}
+
+export function markCompletedSteps(text: string, items: PlanTodoItem[]): number {
+    const doneSteps = extractDoneSteps(text);
+    for (const step of doneSteps) {
+        const item = items.find((t) => t.step === step);
+        if (item) item.completed = true;
+    }
+    return doneSteps.length;
+}
+
+// ── Helpers ──────────────────────────────────────────────────────────────────
+
+export function isAssistantMessage(m: AgentMessage): m is AssistantMessage {
+    return m.role === "assistant" && Array.isArray(m.content);
+}
+
+export function getTextContent(message: AssistantMessage): string {
+    return message.content
+        .filter((block): block is TextContent => block.type === "text")
+        .map((block) => block.text)
+        .join("\n");
+}
