+++ packages/cli/src/extensions/remote-plan-mode.ts
@@ -6,7 +6,7 @@
 
 import { randomUUID } from "node:crypto";
 import type { ExtensionContext } from "@mariozechner/pi-coding-agent";
-import { setPlanModeFromRemote, requestContextClear } from "./plan-mode-toggle.js";
+import { setPlanModeFromRemote, requestContextClear } from "./plan-mode/index.js";
 import type {
     RelayContext,
     PlanModeParams,
